Added -Silent parameter
Changes the default behaviour of signing certificates from the Windows
key provider to not be silent. This means that if a key requires a PIN
to be used then Windows will open a prompt requesting the key. This is
needed to use keys protected by HSM devices like Yubikey which requires
the pin prompt.

The parameter -Silent has been added to the
Add/Set-OpenAuthenticodeSignature cmdlets to bring back the original
behaviour if it is desired.